+++ trunk/EQEmuServer/zone/effects.cpp
@@ -498,7 +498,7 @@ bool Client::TrainDiscipline(int32 itemid) {
 	const Item_Struct *item = database.GetItem(itemid);
 	if(item == NULL) {
 		Message(13, "Unable to find the tome you turned in!");
-		LogFile->write(EQEMuLog::Error, "Unable to find turned in tome id %lu\n", itemid);
+		LogFile->write(EQEMuLog::Error, "Unable to find turned in tome id %lu\n", (unsigned long)itemid);
 		return(false);
 	}
